+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/LanguageRegistrationProcessor.java
@@ -73,7 +73,6 @@ import javax.tools.StandardLocation;
 import com.oracle.truffle.api.TruffleLanguage;
 import com.oracle.truffle.api.TruffleLanguage.Registration;
 import com.oracle.truffle.dsl.processor.java.ElementUtils;
-import java.util.Map;
 
 @SupportedAnnotationTypes("com.oracle.truffle.api.TruffleLanguage.Registration")
 public final class LanguageRegistrationProcessor extends AbstractProcessor {
@@ -138,19 +137,12 @@ public final class LanguageRegistrationProcessor extends AbstractProcessor {
             p.setProperty(prefix + "internal", Boolean.toString(annotation.internal()));
 
             int serviceCounter = 0;
+            TypeMirror registrationType = ProcessorContext.getInstance().getType(Registration.class);
             for (AnnotationMirror anno : l.getAnnotationMirrors()) {
-                final String annoName = anno.getAnnotationType().asElement().toString();
-                if (Registration.class.getCanonicalName().equals(annoName)) {
-                    for (Map.Entry<? extends ExecutableElement, ? extends AnnotationValue> entry : anno.getElementValues().entrySet()) {
-                        final CharSequence attrName = entry.getKey().getSimpleName();
-                        if ("services".contentEquals(attrName)) {
-                            AnnotationValue attrValue = entry.getValue();
-                            List<?> classes = (List<?>) attrValue.getValue();
-                            for (Object clazz : classes) {
-                                AnnotationValue clazzValue = (AnnotationValue) clazz;
-                                p.setProperty(prefix + "service" + serviceCounter++, clazzValue.getValue().toString());
-                            }
-                        }
+                TypeMirror annoType = anno.getAnnotationType();
+                if (ElementUtils.typeEquals(registrationType, annoType)) {
+                    for (TypeMirror serviceTypeMirror : ElementUtils.getAnnotationValueList(TypeMirror.class, anno, "services")) {
+                        p.setProperty(prefix + "service" + serviceCounter++, ElementUtils.getQualifiedName(serviceTypeMirror));
                     }
                 }
             }