+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/LanguageCache.java
@@ -65,6 +65,7 @@ import com.oracle.truffle.api.TruffleLanguage;
 import com.oracle.truffle.api.TruffleLanguage.ContextPolicy;
 import com.oracle.truffle.api.TruffleLanguage.Registration;
 import com.oracle.truffle.api.TruffleOptions;
+import java.util.Arrays;
 
 /**
  * Ahead-of-time initialization. If the JVM is started with {@link TruffleOptions#AOT}, it populates
@@ -93,6 +94,7 @@ final class LanguageCache implements Comparable<LanguageCache> {
     private String languageHome;
     private volatile ContextPolicy policy;
     private volatile Class<? extends TruffleLanguage<?>> languageClass;
+    private volatile Collection<Class<?>> serviceClasses;
 
     private LanguageCache(String id, String prefix, Properties info, ClassLoader loader, String url) {
         this.loader = loader;
@@ -136,6 +138,7 @@ final class LanguageCache implements Comparable<LanguageCache> {
         if (TruffleOptions.AOT) {
             initializeLanguageClass();
             assert languageClass != null;
+            assert serviceClasses != null;
             assert policy != null;
         }
         this.globalInstance = null;
@@ -155,7 +158,7 @@ final class LanguageCache implements Comparable<LanguageCache> {
 
     @SuppressWarnings("unchecked")
     LanguageCache(String id, String name, String implementationName, String version, boolean interactive, boolean internal,
-                    TruffleLanguage<?> instance, String... services) {
+                    TruffleLanguage<?> instance, Class<?>... services) {
         this.id = id;
         this.className = instance.getClass().getName();
         this.mimeTypes = Collections.emptySet();
@@ -175,9 +178,13 @@ final class LanguageCache implements Comparable<LanguageCache> {
         this.globalInstance = instance;
         if (services.length == 0) {
             this.services = Collections.emptySet();
+            this.serviceClasses = Collections.emptySet();
         } else {
+            this.serviceClasses = Collections.unmodifiableCollection(Arrays.<Class<?>> asList(services));
             this.services = new TreeSet<>();
-            Collections.addAll(this.services, services);
+            for (Class<?> service : services) {
+                this.services.add(service.getName());
+            }
         }
     }
 
@@ -422,12 +429,19 @@ final class LanguageCache implements Comparable<LanguageCache> {
         return policy;
     }
 
-    String[] serices() {
-        return services.toArray(new String[services.size()]);
+    Collection<Class<?>> getServices() {
+        initializeLanguageClass();
+        assert serviceClasses != null;
+        return serviceClasses;
     }
 
     boolean supportsService(Class<?> clazz) {
-        return services.contains(clazz.getName()) || services.contains(clazz.getCanonicalName());
+        for (Class<?> serviceClass : getServices()) {
+            if (clazz.equals(serviceClass)) {
+                return true;
+            }
+        }
+        return false;
     }
 
     @SuppressWarnings("unchecked")
@@ -449,6 +463,16 @@ final class LanguageCache implements Comparable<LanguageCache> {
                     } catch (ClassNotFoundException e) {
                         throw new IllegalStateException("Cannot load language " + name + ". Language implementation class " + className + " failed to load.", e);
                     }
+                    assert serviceClasses == null;
+                    List<Class<?>> serviceClassesCollector = new ArrayList<>(services.size());
+                    for (String serviceClassName : services) {
+                        try {
+                            serviceClassesCollector.add(Class.forName(serviceClassName, true, loader));
+                        } catch (ClassNotFoundException e) {
+                            throw new IllegalStateException("Cannot load language " + name + ". Language service class " + serviceClassName + " failed to load.", e);
+                        }
+                    }
+                    serviceClasses = Collections.unmodifiableCollection(serviceClassesCollector);
                 }
             }
         }